+++ packages/genesys-spark-components/src/components/stable/gux-dropdown-multi/tests/gux-dropdown-multi.e2e.ts
@@ -1,9 +1,9 @@
 import { E2EElement, E2EPage } from '@stencil/core/testing';
 import { newSparkE2EPage, a11yCheck } from '../../../../test/e2eTestUtils';
 
-describe('gux-dropdown-multi-beta', () => {
+describe('gux-dropdown-multi', () => {
   const html = `
-    <gux-dropdown-multi-beta lang="en">
+    <gux-dropdown-multi lang="en">
       <gux-listbox-multi aria-label="Animals">
         <gux-option-multi value="a">Ant</gux-option-multi>
         <gux-option-multi value="b">Bat</gux-option-multi>
@@ -15,12 +15,12 @@ describe('gux-dropdown-multi-beta', () => {
         <gux-option-multi value="h">Horse</gux-option-multi>
         <gux-option-multi value="i">Ibis</gux-option-multi>
       </gux-listbox-multi>
-    </gux-dropdown-multi-beta>
+    </gux-dropdown-multi>
   `;
   describe('#render', () => {
     it('renders', async () => {
       const page = await newSparkE2EPage({ html });
-      const element = await page.find('gux-dropdown-multi-beta');
+      const element = await page.find('gux-dropdown-multi');
       expect(element).toHaveAttribute('hydrated');
     });
   });
@@ -50,7 +50,7 @@ describe('gux-dropdown-multi-beta', () => {
       expect(selectedItems.length).toBe(0);
 
       const listboxItems = await page.findAll(
-        'gux-dropdown-multi-beta gux-listbox-multi gux-option-multi'
+        'gux-dropdown-multi gux-listbox-multi gux-option-multi'
       );
       await listboxItems[0].click();
       await page.waitForChanges();
@@ -76,7 +76,7 @@ describe('gux-dropdown-multi-beta', () => {
       await page.waitForChanges();
 
       const listboxItems = await page.findAll(
-        'gux-dropdown-multi-beta gux-listbox-multi gux-option-multi'
+        'gux-dropdown-multi gux-listbox-multi gux-option-multi'
       );
       await listboxItems[0].click();
       await listboxItems[1].click();
@@ -129,9 +129,7 @@ describe('gux-dropdown-multi-beta', () => {
       const dropdownButtonElement = await page.find('pierce/.gux-field');
       await dropdownButtonElement.press('ArrowDown');
 
-      const listbox = await page.find(
-        'gux-dropdown-multi-beta gux-listbox-multi'
-      );
+      const listbox = await page.find('gux-dropdown-multi gux-listbox-multi');
       const focusEl = await page.find(':focus');
       expect(listbox.outerHTML).toContain(focusEl.outerHTML);
     });
@@ -142,11 +140,9 @@ describe('gux-dropdown-multi-beta', () => {
       const dropdownButtonElement = await page.find('pierce/.gux-field');
       await dropdownButtonElement.press('ArrowDown');
 
-      const listbox = await page.find(
-        'gux-dropdown-multi-beta gux-listbox-multi'
-      );
+      const listbox = await page.find('gux-dropdown-multi gux-listbox-multi');
       const listboxItems = await page.findAll(
-        'gux-dropdown-multi-beta gux-listbox-multi gux-option-multi'
+        'gux-dropdown-multi gux-listbox-multi gux-option-multi'
       );
 
       let activeItem = await listbox.findAll('.gux-active');
@@ -167,7 +163,7 @@ describe('gux-dropdown-multi-beta', () => {
       await dropdownButtonElement.press('ArrowDown');
 
       let listboxItems = await page.findAll(
-        'gux-dropdown-multi-beta gux-listbox-multi gux-option-multi'
+        'gux-dropdown-multi gux-listbox-multi gux-option-multi'
       );
       let selectedItems = await page.findAll('.gux-selected');
       expect(selectedItems.length).toBe(0);
@@ -177,7 +173,7 @@ describe('gux-dropdown-multi-beta', () => {
 
       selectedItems = await page.findAll('.gux-selected');
       listboxItems = await page.findAll(
-        'gux-dropdown-multi-beta gux-listbox-multi gux-option-multi'
+        'gux-dropdown-multi gux-listbox-multi gux-option-multi'
       );
 
       expect(selectedItems.length).toBe(1);
@@ -199,42 +195,9 @@ describe('gux-dropdown-multi-beta', () => {
     });
   });
   describe('filter', () => {
-    // remove filterable example in V4 (COMUI-1369)
-    it('filters dropdown contents (filterable)', async () => {
-      const filterableDropdown = `
-      <gux-dropdown-multi-beta filterable lang="en">
-        <gux-listbox-multi aria-label="Animals">
-          <gux-option-multi value="ant">Ant</gux-option-multi>
-          <gux-option-multi value="bear">Bear</gux-option-multi>
-          <gux-option-multi value="bat">Bat</gux-option-multi>
-          <gux-option-multi value="cat">Cat</gux-option-multi>
-          <gux-option-multi value="dog">Dog</gux-option-multi>
-        </gux-listbox-multi>
-      </gux-dropdown-multi-beta>
-    `;
-      const page = await newSparkE2EPage({ html: filterableDropdown });
-      await page.waitForChanges();
-      const dropdownButtonElement = await page.find('pierce/.gux-field');
-      await dropdownButtonElement.click();
-      await page.waitForChanges();
-
-      let listboxItems = await page.findAll(
-        'gux-dropdown-multi-beta gux-listbox-multi gux-option-multi:not(.gux-filtered)'
-      );
-
-      expect(listboxItems.length).toBe(5);
-      await page.keyboard.press('b');
-      await page.waitForChanges();
-
-      listboxItems = await page.findAll(
-        'gux-dropdown-multi-beta gux-listbox-multi gux-option-multi:not(.gux-filtered)'
-      );
-      expect(listboxItems.length).toBe(2);
-      expect(listboxItems[0].textContent).toEqual('Bear');
-    });
     it('filters dropdown contents (filter-type starts-with)', async () => {
       const filterableDropdown = `
-      <gux-dropdown-multi-beta filter-type="starts-with" lang="en">
+      <gux-dropdown-multi filter-type="starts-with" lang="en">
         <gux-listbox-multi aria-label="Animals">
           <gux-option-multi value="ant">Ant</gux-option-multi>
           <gux-option-multi value="bear">Bear</gux-option-multi>
@@ -242,7 +205,7 @@ describe('gux-dropdown-multi-beta', () => {
           <gux-option-multi value="cat">Cat</gux-option-multi>
           <gux-option-multi value="dog">Dog</gux-option-multi>
         </gux-listbox-multi>
-      </gux-dropdown-multi-beta>
+      </gux-dropdown-multi>
     `;
       const page = await newSparkE2EPage({ html: filterableDropdown });
       await page.waitForChanges();
@@ -251,7 +214,7 @@ describe('gux-dropdown-multi-beta', () => {
       await page.waitForChanges();
 
       let listboxItems = await page.findAll(
-        'gux-dropdown-multi-beta gux-listbox-multi gux-option-multi:not(.gux-filtered)'
+        'gux-dropdown-multi gux-listbox-multi gux-option-multi:not(.gux-filtered)'
       );
 
       expect(listboxItems.length).toBe(5);
@@ -259,14 +222,14 @@ describe('gux-dropdown-multi-beta', () => {
       await page.waitForChanges();
 
       listboxItems = await page.findAll(
-        'gux-dropdown-multi-beta gux-listbox-multi gux-option-multi:not(.gux-filtered)'
+        'gux-dropdown-multi gux-listbox-multi gux-option-multi:not(.gux-filtered)'
       );
       expect(listboxItems.length).toBe(2);
       expect(listboxItems[0].textContent).toEqual('Bear');
     });
     it('does not filter dropdown contents (filter-type custom)', async () => {
       const filterableDropdown = `
-      <gux-dropdown-multi-beta filter-type="custom" lang="en">
+      <gux-dropdown-multi filter-type="custom" lang="en">
         <gux-listbox-multi aria-label="Animals">
           <gux-option-multi value="ant">Ant</gux-option-multi>
           <gux-option-multi value="bear">Bear</gux-option-multi>
@@ -274,7 +237,7 @@ describe('gux-dropdown-multi-beta', () => {
           <gux-option-multi value="cat">Cat</gux-option-multi>
           <gux-option-multi value="dog">Dog</gux-option-multi>
         </gux-listbox-multi>
-      </gux-dropdown-multi-beta>
+      </gux-dropdown-multi>
     `;
       const page = await newSparkE2EPage({ html: filterableDropdown });
       await page.waitForChanges();
@@ -283,7 +246,7 @@ describe('gux-dropdown-multi-beta', () => {
       await page.waitForChanges();
 
       let listboxItems = await page.findAll(
-        'gux-dropdown-multi-beta gux-listbox-multi gux-option-multi:not(.gux-filtered)'
+        'gux-dropdown-multi gux-listbox-multi gux-option-multi:not(.gux-filtered)'
       );
 
       expect(listboxItems.length).toBe(5);
@@ -291,7 +254,7 @@ describe('gux-dropdown-multi-beta', () => {
       await page.waitForChanges();
 
       listboxItems = await page.findAll(
-        'gux-dropdown-multi-beta gux-listbox-multi gux-option-multi:not(.gux-filtered)'
+        'gux-dropdown-multi gux-listbox-multi gux-option-multi:not(.gux-filtered)'
       );
       expect(listboxItems.length).toBe(5);
       expect(listboxItems[0].textContent).toEqual('Ant');
@@ -356,7 +319,7 @@ async function setupPage(html: string): Promise<{
 }> {
   const page = await newSparkE2EPage({ html });
   await page.waitForChanges();
-  const dropdown = await page.find('gux-dropdown-multi-beta');
+  const dropdown = await page.find('gux-dropdown-multi');
   const listbox = await page.find('gux-listbox-multi');
   const listboxOptions = await page.findAll('gux-option-multi');
   return { page, dropdown, listbox, listboxOptions };
@@ -404,23 +367,23 @@ async function addNewCustomOption(page: E2EPage): Promise<void> {
 }
 
 const creatableDropdown = `
-<gux-dropdown-multi-beta lang="en">
+<gux-dropdown-multi lang="en">
   <gux-listbox-multi aria-label="Animals">
     <gux-create-option slot="create"></gux-create-option>
     <gux-option-multi value="ant">Ant</gux-option-multi>
     <gux-option-multi value="bear">Bear</gux-option-multi>
     <gux-option-multi value="cat">Cat</gux-option-multi>
   </gux-listbox-multi>
-</gux-dropdown-multi-beta>
+</gux-dropdown-multi>
 `;
 
 const valueSetDropdown = `
-<gux-dropdown-multi-beta lang="en" value="newoption">
+<gux-dropdown-multi lang="en" value="newoption">
   <gux-listbox-multi aria-label="Animals">
     <gux-create-option slot="create"></gux-create-option>
     <gux-option-multi value="ant">Ant</gux-option-multi>
     <gux-option-multi value="bear">Bear</gux-option-multi>
     <gux-option-multi value="cat">Cat</gux-option-multi>
   </gux-listbox-multi>
-</gux-dropdown-multi-beta>
+</gux-dropdown-multi>
 `;