+++ packages/genesys-spark-components/src/components/stable/gux-dropdown-multi/tests/gux-dropdown-multi.spec.ts
@@ -2,18 +2,18 @@ import { newSpecPage } from '@stencil/core/testing';
 import MutationObserver from 'mutation-observer';
 
 import { GuxDropdownMulti } from '../gux-dropdown-multi';
-import { GuxListboxMulti } from '../../../stable/gux-listbox-multi/gux-listbox-multi';
-import { GuxOptionMulti } from '../../../stable/gux-listbox-multi/gux-option-multi/gux-option-multi';
+import { GuxListboxMulti } from '../../gux-listbox-multi/gux-listbox-multi';
+import { GuxOptionMulti } from '../../gux-listbox-multi/gux-option-multi/gux-option-multi';
 
 const components = [GuxDropdownMulti, GuxListboxMulti, GuxOptionMulti];
 const html = `
-<gux-dropdown-multi-beta>
+<gux-dropdown-multi>
   <gux-listbox-multi aria-label="Animals">
     <gux-option-multi value="a" disabled>Ant</gux-option-multi>
     <gux-option-multi value="b">Bat</gux-option-multi>
     <gux-option-multi value="c">Cat</gux-option-multi>
   </gux-listbox-multi>
-</gux-dropdown-multi-beta>
+</gux-dropdown-multi>
 `;
 const language = 'en';